requirements.txt
asgiref==3.8.1
boto3==1.20.26
botocore==1.23.54
certifi==2025.4.26
charset-normalizer==3.4.2
Django==4.2
django-anymail==13.0
django-cors-headers==3.14.0
django-jazzmin==3.0.1
django-storages==1.12.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.10
environs==14.2.0
idna==3.10
inflection==0.5.1
jmespath==0.10.0
marshmallow==4.0.0
packaging==25.0
PyJWT==2.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.5.2
setuptools==80.8.0
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.3
stripe==12.2.0
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==1.26.20
